# docs/api/openapi.yaml
components:
  schemas:
...
    OperationRecord:
      additionalProperties: false
      properties:
        actor_user_id:
          anyOf:
          - format: int64
            type: integer
          - type: 'null'
        created_at:
          type: string
        id:
          format: int64
          type: integer
        method:
          type: string
        path:
          type: string
        status:
          format: int64
          type: integer
      required:
      - id
      - method
      - path
      - status
      - created_at
      type: object
    OperationRecordCountBucket:
      additionalProperties: false
      properties:
        count:
          format: int64
          type: integer
        key:
          type: string
      required:
      - key
      - count
      type: object
...
    OperationRecordPathBucket:
      additionalProperties: false
      properties:
        count:
          format: int64
          type: integer
        error_count:
          format: int64
          type: integer
        last_seen_at:
          anyOf:
          - type: string
          - type: 'null'
        path:
          type: string
      required:
      - path
      - count
      - error_count
      type: object
...
    OperationRecordSummary:
      additionalProperties: false
      properties:
        by_method:
          items:
            $ref: '#/components/schemas/OperationRecordCountBucket'
          type: array
        by_status_class:
          items:
            $ref: '#/components/schemas/OperationRecordCountBucket'
          type: array
        client_error_count:
          format: int64
          type: integer
        generated_at:
          type: string
        other_count:
          format: int64
          type: integer
        redirect_count:
          format: int64
          type: integer
        server_error_count:
          format: int64
          type: integer
        success_count:
          format: int64
          type: integer
        top_limit:
          format: int64
          type: integer
        top_paths:
          items:
            $ref: '#/components/schemas/OperationRecordPathBucket'
          type: array
        total_count:
          format: int64
          type: integer
      required:
      - generated_at
      - total_count
      - success_count
      - redirect_count
      - client_error_count
      - server_error_count
      - other_count
      - top_limit
      - by_method
      - by_status_class
      - top_paths
      type: object
...
    TrafficProbeAlertActionResult:
      additionalProperties: false
      properties:
        updated:
          type: boolean
      required:
      - updated
      type: object
...
    VersionPackageActionRequest:
      additionalProperties: false
      properties:
        reason:
          anyOf:
          - type: string
          - type: 'null'
      required: []
      type: object
    VersionPackageActionResult:
      additionalProperties: false
      properties:
        event_id:
          format: int64
          type: integer
        package:
          $ref: '#/components/schemas/VersionPackageEntry'
        previous_active_id:
          anyOf:
          - format: int64
            type: integer
          - type: 'null'
      required:
      - event_id
      - package
      type: object
    VersionPackageEntry:
      additionalProperties: false
      properties:
        created_at:
          type: string
        id:
          format: int64
          type: integer
        manifest:
          additionalProperties: true
          description: 任意 JSON 值
        published_at:
          anyOf:
          - type: string
          - type: 'null'
        retired_at:
          anyOf:
          - type: string
          - type: 'null'
        status:
          type: string
        version_code:
          type: string
        version_name:
          type: string
      required:
      - id
      - version_name
      - version_code
      - manifest
      - status
      - created_at
      type: object
...
info:
  description: 由 Rust route registry 生成的当前有效 API 契约，禁止手写漂移。
  title: Aoi[葵] API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/system/operation-records:
    get:
      operationId: system.operation-records
      parameters:
      - description: 按 HTTP method 精确过滤
        in: query
        name: method
        required: false
        schema:
          type: string
      - description: 按 route 模板精确过滤，不包含 query/fragment
        in: query
        name: path
        required: false
        schema:
          type: string
      - description: 按 HTTP 状态码过滤
        in: query
        name: status
        required: false
        schema:
          type: integer
      - description: 按操作者用户 ID 过滤
        in: query
        name: actor_user_id
        required: false
        schema:
          type: integer
      - description: 按创建时间起点过滤，RFC3339
        in: query
        name: created_from
        required: false
        schema:
          type: string
      - description: 按创建时间终点过滤，RFC3339
        in: query
        name: created_to
        required: false
        schema:
          type: string
      - description: 返回条数，范围 1..200
        in: query
        name: limit
        required: false
        schema:
          type: integer
      - description: 分页偏移，范围 0..10000
        in: query
        name: offset
        required: false
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/OperationRecord'
                type: array
          description: 成功
      summary: 读取操作记录
      tags:
      - System
      x-console-access: permission
      x-console-permission: operation_record:read
      x-console-product-code: console
      x-console-scope: platform
...
  /api/v1/system/operation-records/summary:
    get:
      operationId: system.operation-records.summary
      parameters:
      - description: 按 HTTP method 精确过滤
        in: query
        name: method
        required: false
        schema:
          type: string
      - description: 按 route 模板精确过滤，不包含 query/fragment
        in: query
        name: path
        required: false
        schema:
          type: string
      - description: 按 HTTP 状态码过滤
        in: query
        name: status
        required: false
        schema:
          type: integer
      - description: 按操作者用户 ID 过滤
        in: query
        name: actor_user_id
        required: false
        schema:
          type: integer
      - description: 按创建时间起点过滤，RFC3339
        in: query
        name: created_from
        required: false
        schema:
          type: string
      - description: 按创建时间终点过滤，RFC3339
        in: query
        name: created_to
        required: false
        schema:
          type: string
      - description: top route 返回条数，范围 1..50
        in: query
        name: top_limit
        required: false
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationRecordSummary'
          description: 成功
      summary: 读取操作记录审计汇总
      tags:
      - System
      x-console-access: permission
      x-console-permission: operation_record:read
      x-console-product-code: console
      x-console-scope: platform
...
  /api/v1/system/traffic-probes/alerts/{id}/resolve:
    post:
      operationId: system.traffic-probes.alerts.resolve
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrafficProbeAlertActionResult'
          description: 成功
      summary: 恢复流量探针告警
      tags:
      - System Traffic Probe
      x-console-access: permission
      x-console-permission: traffic_probe:write
      x-console-product-code: console
      x-console-scope: platform
...
  /api/v1/system/version-packages/{id}/publish:
    post:
      operationId: system.version-packages.publish
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionPackageActionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionPackageActionResult'
          description: 成功
      summary: 发布版本包为 active
      tags:
      - System Version
      x-console-access: permission
      x-console-permission: version_package:write
      x-console-product-code: console
      x-console-scope: platform
  /api/v1/system/version-packages/{id}/rollback:
    post:
      operationId: system.version-packages.rollback
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionPackageActionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionPackageActionResult'
          description: 成功
      summary: 回滚到指定版本包
      tags:
      - System Version
      x-console-access: permission
      x-console-permission: version_package:write
      x-console-product-code: console
      x-console-scope: platform
...